Use double instead of float for decimal positions
A position ranges from 0 to Integer.MAX_VALUE, but floats only have a 23
bit mantissa. Values over 2^23 are rounded to a nearby value.

Solve this by using double instead of float when we need to represent a
position with a fractional component. A double has a mantissa of 52 bit
and can therefore represent all 2^32 possible int values.

Bug: 134858960
Test: ./gradlew viewpager2:cC

+++ viewpager2/src/androidTest/java/androidx/viewpager2/widget/BaseTest.kt
@@ -413,7 +413,7 @@ open class BaseTest {
         return latch
     }
 
-    fun ViewPager2.addWaitForDistanceToTarget(target: Int, distance: Float): CountDownLatch {
+    fun ViewPager2.addWaitForDistanceToTarget(target: Int, distance: Double): CountDownLatch {
         val latch = CountDownLatch(1)
 
         registerOnPageChangeCallback(object : ViewPager2.OnPageChangeCallback() {
@@ -422,7 +422,7 @@ open class BaseTest {
                 positionOffset: Float,
                 positionOffsetPixels: Int
             ) {
-                if (abs(target - position - positionOffset) <= distance) {
+                if (abs(target - position - positionOffset.toDouble()) <= distance) {
                     latch.countDown()
                     post { unregisterOnPageChangeCallback(this) }
                 }
